+++ plugins/test/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @livekit/agents-plugins-test
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`4627059d74346ac07282dcc8d02a68b1355b6741`](https://github.com/livekit/agents-js/commit/4627059d74346ac07282dcc8d02a68b1355b6741)]:
+  - @livekit/agents@1.0.8
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`b1c835817e0abec4cdc468cbd4e37fdd4635dde5`](https://github.com/livekit/agents-js/commit/b1c835817e0abec4cdc468cbd4e37fdd4635dde5), [`5f39a0c68d407fca2b9f46feb438dad47939d4fb`](https://github.com/livekit/agents-js/commit/5f39a0c68d407fca2b9f46feb438dad47939d4fb), [`e43a6a190b2e48ea0674c78e0ee7ed7cd9bdc5b9`](https://github.com/livekit/agents-js/commit/e43a6a190b2e48ea0674c78e0ee7ed7cd9bdc5b9), [`fd484279234edd83d7f1c0efb536f2a5a73c2846`](https://github.com/livekit/agents-js/commit/fd484279234edd83d7f1c0efb536f2a5a73c2846), [`435aade03993ad04db07d72114b8beaca1ffdc26`](https://github.com/livekit/agents-js/commit/435aade03993ad04db07d72114b8beaca1ffdc26), [`d162565191c7d42f087cecbd07ab92af6bbfaef8`](https://github.com/livekit/agents-js/commit/d162565191c7d42f087cecbd07ab92af6bbfaef8)]:
+  - @livekit/agents@1.0.7
+
 ## 1.0.6
 
 ### Patch Changes